requirements.txt
feedparser==6.0.11
requests==2.32.3
pyyaml==6.0.2
openai==1.54.3
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.34.0
